nvdaHelper: Always compile with level 3 warnings (/W3) and always treet all warnings as errors (/WX). At the same time, clearly state any automatic type conversions the compiler flagged as questionable by using static_cast. In fact the goal should be to remove all c-style casts and only use const_cast/static_cast as they are safer. Use size_t in a few places where we are clearly managing a size, to stop sign/unsign mismatch errors. Note that we cannot move to /Wall as minHook contains code that does not pass.

+++ nvdaHelper/common/winIPCUtils.h
@@ -26,7 +26,7 @@ extern "C" {
  * @param buf address of allocated memory that could hold cch characters where the port string should be written to.
  * @param cch the size of buf in characters
  */
-int generateDesktopSpecificNamespace(wchar_t* buf, int cch);
+size_t generateDesktopSpecificNamespace(wchar_t* buf, size_t cch);
 
 #ifdef __cplusplus
 }
